Refactor into additional modules
